+++ main.py
@@ -1,0 +1,3 @@
+from tools import init_pooling
+
+init_pooling()
