+++ tools/__init__.py
@@ -1,3 +1,1 @@
-from os_tools import get_secret, BASE_DIR
-from logger import get_logger
-# from bot import init_external_webhook, init_telegram_bot_webhook, send_message
+from tools.bot import init_webhook, init_pooling
